refactor: move internal classes and modules to the core package to hide them from the user and improve documentation
