+++ docs/api/timeeval.core.rst
@@ -1,0 +1,37 @@
+timeeval.core package
+=====================
+
+.. warning::
+    This package contains TimeEval-internal classes and functions.
+
+timeeval.core.experiments.Experiment
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. autoclass:: timeeval.core.experiments.Experiment
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+timeeval.core.experiments.Experiments
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. autoclass:: timeeval.core.experiments.Experiments
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+timeeval.core.remote.Remote
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. autoclass:: timeeval.core.remote.Remote
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+timeeval.core.times.Times
+^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. autoclass:: timeeval.core.times.Times
+   :members:
+   :undoc-members:
+   :show-inheritance:
